// src/types/Person.ts
import { Planet } from "./Planet";

export interface Person {
  name: string;
  height: string;
  mass: string;
  hair_color: string;
  skin_color: string;
  eye_color: string;
  birth_year: string;
  gender: string;
  homeworld: string;
  homeworld_full: Planet;
  films: string[];
  species: string[];
  species_names: string[];
  vehicles: string[];
  starships: string[];
  created: string;
  edited: string;
  url: string;
}